Written programs to solve Small_factorials https://www.spoj.com/problems/FCTRL2/
Wrote a python program to calculate the factorial. It was staright forward.

Wrote a cpp rpgram to calculate the factorial. I had to create a function to store large numbers as part of a integer array. I also wrote a function to multiply numbers across an array. I learned a lot.